Move Tester into a separate file.
